db fix

+++ db/migrate/20230703112600_create_posts.rb
@@ -1,7 +1,7 @@
 class CreatePosts < ActiveRecord::Migration[6.1]
   def change
     create_table :posts do |t|
-      t.references :customer, type: :integer, foreign_key: true, null: false
+      t.references :customer, type: :bigint, foreign_key: true, null: false
       t.string :post_name,          null: false
       t.text :description,          null: false
       t.text :material,             null: false
