+++ db/migrate/20230703121547_create_favorites.rb
@@ -1,8 +1,8 @@
 class CreateFavorites < ActiveRecord::Migration[6.1]
   def change
     create_table :favorites do |t|
-      t.references :customer, type: :integer, foreign_key: true, null: false
-      t.references :post, type: :integer, foreign_key: true, null: false
+      t.references :customer, type: :bigint, foreign_key: true, null: false
+      t.references :post, type: :bigint, foreign_key: true, null: false
 
       t.timestamps
     end
